backend/requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
langchain
langchain_community
langchain-google-genai
langchain-groq
google-generativeai
faiss-cpu
python-dotenv
fpdf2
Pillow
pydantic